created seperate function for ce and ent
